+++ day-7/projectpilot/api/db.json
@@ -732,7 +732,7 @@
     },
     {
       "id": "74",
-      "name": "Armstrong, Crona and Ziemann",
+      "name": "HOLA",
       "description": "Switchable bifurcated task-force. Hic sed at et id et et qui quis dolores.",
       "imageUrl": "/assets/placeimg_500_300_arch6.jpg",
       "contractTypeId": 2,
@@ -942,13 +942,13 @@
     },
     {
       "id": "95",
-      "name": "Armstrong - Hands",
+      "name": "HOLA",
       "description": "Persistent 3rd generation capability. Sint omnis asperiores voluptas sapiente delectus blanditiis earum ut voluptatem.",
       "imageUrl": "/assets/placeimg_500_300_arch6.jpg",
       "contractTypeId": 2,
       "contractSignedOn": "2010-09-24T21:02:41.601Z",
       "budget": 38853,
-      "isActive": false
+      "isActive": true
     },
     {
       "id": "97",
