+++ day-7/projectpilot/src/projects/ProjectsPage.tsx
@@ -1,61 +1,26 @@
-import { useState, useEffect } from 'react';
-import { projectAPI } from './projectAPI';
+import { useProjects } from './projectHooks';
 import ProjectList from "./ProjectList";
-import { Project } from './Project';
 
 function ProjectsPage() {
-  // Hook de estado para la lista de proyectos, inicializado con MOCK_PROJECTS (tipo Project[])
-  const [projects, setProjects] = useState<Project[]>([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | undefined>(undefined);
-  const [currentPage, setCurrentPage] = useState(1);
+  const {
+    projects,
+    loading,
+    error,
+    setCurrentPage,
+    saveProject,
+    saving,
+    //savingError
+  } = useProjects();
 
   const handleMoreClick = () => {
-    setCurrentPage((currentPage) => currentPage + 1)
+    setCurrentPage((currentPage) => currentPage + 1);
   }
 
-  useEffect(() => {
-    async function loadProjects() {
-      setLoading(true);
-      try {
-        const data = await projectAPI.get(currentPage);
-        if(currentPage === 1){
-          setProjects(data);
-        } else {
-          setProjects((projects) => [...projects, ...data]); //concatena los proyectos del state con los nuevos gracias a la paginacion
-        }
-      } catch (e) {
-        if (e instanceof Error) {
-          setError(e.message);
-        } else {
-          setError("Unknown error");
-        }
-      } finally {
-        setLoading(false);
-      }
-    }
-    loadProjects();
-    //al momento de modificar el valor de currentPage, se recarga useEffect
-  }, [currentPage]);
-
- const saveProject = (project: Project) => {
-  //Especificar que p es de tipo Project
-  //Crear nueva lista de proyectos, reemplazando el proyecto que tiene el mismo id
-  projectAPI
-    .put(project)
-    .then((updatedProject) => {
-      let updatedProjects = projects.map((p) => {
-        return p.id === project.id ? new Project(updatedProject) : p;
-      });
-      setProjects(updatedProjects);
-    })
-    .catch((e) => {
-      setError(e.message);
-    })
-  };
   return (
     <>
       <h1>Projects</h1>
+
+      {saving && <span className="toast">Saving...</span>}
 
       {error && (
         <div className="row">
